# docs/openapi/agent-platform-openapi.json
{
  "components": {
    "schemas": {
...
      "ApiToolCenterItemResponse": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "last_updated_at": {
            "default": "",
            "title": "Last Updated At",
            "type": "string"
          },
          "managed": {
            "title": "Managed",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "read_only": {
            "title": "Read Only",
            "type": "boolean"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "source_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Code"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "string"
          },
          "tool_type": {
            "title": "Tool Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "tool_id",
          "name",
          "description",
          "tool_type",
          "source_type",
          "managed",
          "read_only",
          "archived"
        ],
        "title": "ApiToolCenterItemResponse",
        "type": "object"
      },
      "ApiToolCenterListResponse": {
        "properties": {
          "include_archived": {
            "title": "Include Archived",
            "type": "boolean"
          },
          "include_builtin": {
            "title": "Include Builtin",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiToolCenterItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "include_archived",
          "include_builtin",
          "items"
        ],
        "title": "ApiToolCenterListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ToolCenterCreateRequest": {
        "properties": {
          "default_requires_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Requires Approval"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "enable_parallel_execution": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Parallel Execution"
          },
          "npm_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npm Requirements"
          },
          "pip_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pip Requirements"
          },
          "return_char_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Char Limit"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_code": {
            "title": "Source Code",
            "type": "string"
          },
          "source_type": {
            "default": "python",
            "title": "Source Type",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "slug",
          "source_code"
        ],
        "title": "ToolCenterCreateRequest",
        "type": "object"
      },
      "ToolCenterUpdateRequest": {
        "properties": {
          "default_requires_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Requires Approval"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "enable_parallel_execution": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Parallel Execution"
          },
          "npm_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npm Requirements"
          },
          "pip_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pip Requirements"
          },
          "return_char_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Char Limit"
          },
          "source_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Code"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "ToolCenterUpdateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Provides versioned API routes for Agent Platform runtime/control/test orchestration. Designed for backend-first API consumption and ADE frontend integration.",
    "summary": "Runtime and control APIs for ADE and local Agent Platform workflows",
    "title": "Agent Platform API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/platform/tool-center/tools": {
      "get": {
        "operationId": "api_tool_center_list_tools_api_v1_platform_tool_center_tools_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_builtin",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Builtin",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_source",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Source",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "title": "Search",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tool Center entries",
        "tags": [
          "Tool Center"
        ]
      },
      "post": {
        "operationId": "api_tool_center_create_tool_api_v1_platform_tool_center_tools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCenterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v1/platform/tool-center/tools/{slug}": {
      "get": {
        "operationId": "api_tool_center_get_tool_api_v1_platform_tool_center_tools__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_source",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Source",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tool Center managed custom tool",
        "tags": [
          "Tool Center"
        ]
      },
      "patch": {
        "operationId": "api_tool_center_update_tool_api_v1_platform_tool_center_tools__slug__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCenterUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v1/platform/tool-center/tools/{slug}/archive": {
      "post": {
        "operationId": "api_tool_center_archive_tool_api_v1_platform_tool_center_tools__slug__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v1/platform/tool-center/tools/{slug}/purge": {
      "delete": {
        "operationId": "api_tool_center_purge_tool_api_v1_platform_tool_center_tools__slug__purge_delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge archived managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v1/platform/tool-center/tools/{slug}/restore": {
      "post": {
        "operationId": "api_tool_center_restore_tool_api_v1_platform_tool_center_tools__slug__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore archived managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Agent Platform API local",
      "url": "http://127.0.0.1:8284"
    }
  ],
...
}